+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 
 ### Fixed
 
+-   Fix private appdata issue (#344)
+
 ## [0.10.1](https://github.com/ansys/pymechanical/releases/tag/v0.10.1) - August 8 2023
 
 ### Changed